refactor: module structure for grammar / spell checking
