+++ csaf-rs/src/validations/mod.rs
@@ -137,5 +137,6 @@ pub mod test_6_3_11;
 pub mod test_6_3_12;
 pub mod test_6_3_14;
 pub mod test_6_3_15;
+pub mod test_6_3_16;
 pub mod test_6_3_18;
 pub mod test_6_3_20;
