+++ csaf-rs/src/validations/utils/text_check/spell/mock_spell.rs
@@ -2,9 +2,8 @@
 //!
 //! This module provides a basic English spell-checker backed by a static word list.
 
-use crate::validations::utils::text_check::checkers::utils::tokenize_words;
-use crate::validations::utils::text_check::checkers::{TemporaryTextCheckQuality, TextChecker};
-use crate::validations::utils::text_check::{TextCheckFinding, TextCheckKind};
+use crate::validations::utils::text_check::utils::{TemporaryTextCheckQuality, tokenize_words};
+use crate::validations::utils::text_check::{TextCheckFinding, TextCheckKind, TextChecker};
 use std::collections::HashSet;
 
 /// A mock spell-checker for English text.
