--- csaf-rs/src/validations/utils/text_check/tests/mod.rs
@@ -1,3 +1,0 @@
-mod integration_tests;
-mod unit_tests;
-mod utils;
